+++ packages/dm-core/src/hooks/useLocalStorage.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
 
 export function useLocalStorage<T>(
   key: string,
@@ -14,16 +14,9 @@ export function useLocalStorage<T>(
     }
   })
 
-  const setValue = (value: T | ((val: T) => T)) => {
-    try {
-      const valueToStore =
-        value instanceof Function ? value(storedValue) : value
-      setStoredValue(valueToStore)
-      window.localStorage.setItem(key, JSON.stringify(valueToStore))
-    } catch (error) {
-      console.log(error)
-    }
-  }
+  useEffect(() => {
+    window.localStorage.setItem(key, JSON.stringify(storedValue))
+  }, [storedValue])
 
-  return [storedValue, setValue]
+  return [storedValue, setStoredValue]
 }