+++ packages/dm-core/src/types.ts
@@ -100,6 +100,14 @@ export type TJob = {
   ended?: string
   outputTarget?: string
   referenceTarget?: string
+  schedule?: TSchedule
+}
+
+export type TSchedule = {
+  type: string
+  cron: string
+  startDate: string
+  endDate: string
 }
 
 export type TJobWithRunner = TJob & {
@@ -165,6 +173,7 @@ export type TUserIdMapping = { userId: string; username: string }
 export type TViewConfig = {
   type: string
   scope?: string
+  resolve?: boolean
   label?: string
   eds_icon?: string
   roles?: string[]